Remove collector subpackage
It has been removed upstream in [1]. According to the commit message,
i understand that the functionality has been included in graph
subpackage so i'm obsoleting -collector in -graph to ensure
upgrade path removes old -collector.

Note that this may break deployment tools which deploy -collector
service and need to be fixed.

[1] http://git.openstack.org/cgit/openstack/vitrage/commit/?id=842f9d6cea1ed85bc4a0f8c30ae9cb4a22765420

Change-Id: Idf3fe105be6f126253dfbeaf936f47ddb352abb8